test: standardize retry logic
